Controllers refactoring
MODIFIED >
Controllers.java
   - Split some of the functionality to their own respective controllers

ADDED >
 UserController.java
    - Responsible for user actions, like logging in/out and creating users etc

 ItemPropertyManagementController.java
    - Responsible for managing Item properties.
